TheEssentialsSweet + core push'd

+++ coreshit/Essentials/asmtypes.h
@@ -5,6 +5,6 @@
 typedef unsigned char byte;
 typedef unsigned short word;
 typedef unsigned int dword;
-typedef unsigned long qword;
+typedef unsigned long long qword;
 
 #endif
